Refactor file paths and imports, and rename components and directories
