--- client/src/features/accounts/AccountTypePicker/index.js
@@ -1,1 +1,0 @@
-export { default } from './AccountTypePicker';
