--- client/src/features/accounts/ActionsMenu/index.js
@@ -1,1 +1,0 @@
-export { default } from './ActionsMenu';
